requirements.txt
aiohttp==3.7.4.post0
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.6.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
cryptography==35.0.0
cycler==0.11.0
discord==1.7.3
discord.py==1.7.3
h11==0.12.0
idna==3.3
kiwisolver==1.3.2
matplotlib==3.4.3
multidict==5.2.0
numpy==1.21.4
outcome==1.1.0
Pillow==8.4.0
pycodestyle==2.8.0
pycparser==2.20
pyOpenSSL==21.0.0
pyparsing==3.0.4
python-dateutil==2.8.2
python-dotenv==0.19.1
requests==2.26.0
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==3.10.0.2
urllib3==1.26.7
wsproto==1.0.0
yarl==1.7.2
